nicer test setup
