+++ bench/start.js
@@ -1,5 +1,11 @@
-function start() {
-  let a = heimdall.start('a');
-}
-
-module.exports = start;
+module.exports = {
+  count: 1000,
+  name: 'Start',
+  setup: function() {
+    var Heimdall = process.Heimdall;
+    var heimdall = new Heimdall();
+  },
+  fn: function() {
+    var a = heimdall.start('a');
+  }
+};
